requirements/base.txt
aiobotocore==2.15.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
botocore==1.35.16
click==8.1.7
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
redis==5.0.8
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
wrapt==1.16.0
yarl==1.11.1
passlib